pyproject.toml
[project]
name = "amazon-ads-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cachetools>=6.2.4",
    "httpx>=0.28.1",
    "pydantic[dotenv]==1.10",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "returns>=0.23.0",
]

[dependency-groups]
dev = ["python-dotenv>=1.2.1"]

[tool.basedpyright]
allowedUntypedLibraries = [
    "returns"
]

typeCheckingMode = "recommended"

strictListInference = true
strictDictionaryInference = true
strictParameterNoneValue = true
enableTypeIgnoreComments = true

reportAny = false
reportExplicitAny = false
reportUntypedFunctionDecorator = false
reportIgnoreCommentWithoutRule = false
reportUnannotatedClassAttribute = false
